deployement scripts for ido
