+++ IDO/.gitignore
@@ -7,3 +7,7 @@ typechain
 #Hardhat files
 cache
 artifacts
+
+#Private
+
+privateKey.json
