+++ IDO/contracts/IDO/IIDO.sol
@@ -7,6 +7,14 @@ import "./enums/Status.sol";
 
 interface IIDO {
 
+    event Locked();
+    event ParamsSet(IERC20 token, uint tokensForSale, IERC20 collateralToken, uint collateralRequired, uint ONEToRaise, uint buyingStartsAt, uint buyingEndsAt, uint vestingStartsAt, uint vestingEndsAt, uint timeToClaim);
+    event TokensPurchased(address user, uint ONEAmount, uint tokenAmount);
+    event TokensClaimed(address user, uint tokenAmount);
+    event CollateralPaid(address user);
+    event CollateralRetrieved(address user);
+
+
     function status() external view returns(STATUS);
 
     receive() external payable;
@@ -21,9 +29,11 @@ interface IIDO {
 
     function withdrawTokens(address to, uint value) external;
 
-    function setParams(IERC20 token, uint tokensForSale, IERC20 collateralToken, uint collateralRequired, uint ONEToRaise, uint buyingStartsAt, uint buyingEndsAt, uint vestingStartsAt, uint vestingEndsAt, uint timeToClaim) external;
+    function setParams(IERC20 token, uint tokensForSale, IERC20 collateralToken, uint collateralRequired, uint ONEToRaise, uint buyingStartsAt, uint buyingEndsAt, uint vestingStartsAt, uint vestingEndsAt, uint timeToClaim, uint maximumTokensPerWallet) external;
 
     function isLockedIn() external view returns(bool);
 
     function lockIn() external;
+
+    function retrieveCollateral() external;
 }
