+++ IDO/hardhat.config.js
@@ -1,4 +1,5 @@
 require("@nomiclabs/hardhat-ethers");
+const privateKey = require("./privateKey")
 
 // This is a sample Hardhat task. To learn how to create your own go to
 // https://hardhat.org/guides/create-task.html
@@ -18,4 +19,10 @@ task("accounts", "Prints the list of accounts", async (taskArgs, hre) => {
  */
 module.exports = {
   solidity: "0.8.4",
+  networks: {
+    Harmony: {
+      url: "https://api.fuzz.fi",
+      accounts: [privateKey]
+    }
+  }
 };
